Rename to detection_datamodule.py
