+++ yolort/datasets/__init__.py
@@ -1,2 +1,2 @@
 # Copyright (c) 2021, Zhiqiang Wang. All Rights Reserved.
-from .pl_datamodule import DetectionDataModule, VOCDetectionDataModule, CocoDetectionDataModule
+from .detection_datamodule import DetectionDataModule, VOCDetectionDataModule, CocoDetectionDataModule
